+++ module_groups/ecs_plb_platform/main.tf
@@ -9,6 +9,7 @@ module "ecs_core_platform" {
   lc_instance_type    = var.lc_instance_type
   max_size            = var.max_size
   min_size            = var.min_size
+  office_ip           = var.office_ip
   health_check_type   = var.health_check_type
   peer_account_id     = var.peer_account_id
   peer_vpc_cidr       = var.peer_vpc_cidr